requirements.txt
torch
torchvision
albumentations
pillow
tqdm
math
einops
functools